+++ plasmapy/formulary/tests/test_dimensionless.py
@@ -1,19 +1,43 @@
+"""Tests for functionality contained in `plasmapy.formulary.dimensionless`."""
+
 import astropy.units as u
 import numpy as np
 import pytest
 
 from plasmapy.formulary.dimensionless import (
     beta,
+    Debye_number,
     Mag_Reynolds,
+    nD_,
     quantum_theta,
     Re_,
     Reynolds_number,
     Rm_,
 )
+from plasmapy.utils.pytest_helpers import assert_can_handle_nparray
+
+Z = 1
 
 B = 1.0 * u.T
+
 n = 5e19 * u.m ** -3
+n_e = Z * 5e19 * u.m ** -3
+
 T = 1e6 * u.K
+T_e = 1e6 * u.K
+
+
+@pytest.mark.parametrize(
+    "alias, parent",
+    [
+        (nD_, Debye_number),
+        (Re_, Reynolds_number),
+        (Rm_, Mag_Reynolds),
+    ],
+)
+def test_parameters_aliases(alias, parent):
+    """Test all aliases defined in parameters.py"""
+    assert alias is parent
 
 
 def test_beta_dimensionless():
@@ -72,8 +96,40 @@ def test_Mag_Reynolds():
         Mag_Reynolds(2.2 * u.kg, L, sigma)
 
 
-def test_dimensionless_aliases():
-    r"""Test all aliases defined in dimensionless.py"""
+def test_Debye_number():
+    r"""Test the Debye_number function in parameters.py."""
 
-    assert Re_ is Reynolds_number
-    assert Rm_ is Mag_Reynolds
+    assert Debye_number(T_e, n_e).unit.is_equivalent(u.dimensionless_unscaled)
+
+    T_e_eV = T_e.to(u.eV, equivalencies=u.temperature_energy())
+    assert np.isclose(Debye_number(T_e, n_e).value, Debye_number(T_e_eV, n_e).value)
+
+    assert np.isclose(Debye_number(1 * u.eV, 1 * u.cm ** -3).value, 1720862385.43342)
+
+    with pytest.warns(u.UnitsWarning):
+        Debye_number(T_e, 4)
+
+    with pytest.raises(ValueError):
+        Debye_number(None, n_e)
+
+    with pytest.raises(u.UnitTypeError):
+        Debye_number(5 * u.m, 5 * u.m ** -3)
+
+    with pytest.raises(u.UnitTypeError):
+        Debye_number(5 * u.K, 5 * u.m ** 3)
+
+    with pytest.raises(ValueError):
+        Debye_number(5j * u.K, 5 * u.cm ** -3)
+
+    Tarr2 = np.array([1, 2]) * u.K
+    narr3 = np.array([1, 2, 3]) * u.m ** -3
+    with pytest.raises(ValueError):
+        Debye_number(Tarr2, narr3)
+
+    with pytest.warns(u.UnitsWarning):
+        assert Debye_number(1.1, 1.1) == Debye_number(1.1 * u.K, 1.1 * u.m ** -3)
+
+    with pytest.warns(u.UnitsWarning):
+        assert Debye_number(1.1 * u.K, 1.1) == Debye_number(1.1, 1.1 * u.m ** -3)
+
+    assert_can_handle_nparray(Debye_number)
